refactor: simplify build and packaging of navs
